Fix read_wav_header return value type
Backported from https://github.com/rs1729/RS

+++ imet/imet4iq.c
@@ -78,7 +78,7 @@ int findstr(char *buff, char *str, int pos) {
 }
 
 static
-float read_wav_header(pcm_t *pcm, FILE *fp) {
+int read_wav_header(pcm_t *pcm, FILE *fp) {
     char txt[4+1] = "\0\0\0\0";
     unsigned char dat[4];
     int byte, p=0;
@@ -1612,4 +1612,3 @@ int main(int argc, char *argv[]) {
 
     return 0;
 }
-
